+++ include/notcurses/notcurses.h
@@ -2387,13 +2387,13 @@ API const char* ncmetric(uintmax_t val, uintmax_t decimal, char* buf,
 
 // Mega, kilo, gigafoo. Use PREFIXSTRLEN + 1.
 static inline const char*
-qprefix(uintmax_t val, unsigned decimal, char* buf, int omitdec){
+qprefix(uintmax_t val, uintmax_t decimal, char* buf, int omitdec){
   return ncmetric(val, decimal, buf, omitdec, 1000, '\0');
 }
 
 // Mibi, kebi, gibibytes. Use BPREFIXSTRLEN + 1.
 static inline const char*
-bprefix(uintmax_t val, unsigned decimal, char* buf, int omitdec){
+bprefix(uintmax_t val, uintmax_t decimal, char* buf, int omitdec){
   return ncmetric(val, decimal, buf, omitdec, 1024, 'i');
 }